Fix/41 naming condition regex (#48)
* Updating for telemetry

* feat: Update validation condition for resource name in variables.tf

The validation condition for the `name` variable in `variables.tf` has been updated to allow for a wider range of characters. The new condition allows for lowercase and uppercase letters, numbers, periods, underscores, hyphens, and parentheses. The name must be between 1 and 254 characters long. This change was made to ensure that the resource name can accommodate a wider range of naming conventions.